[models] refactor: move CPU init helpers to utils
